Combine fill_db with parse and rename to parse
